+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix COM object leaks by releasing all Outlook COM references after use
+
 ## v1.1.0.71 - 2026-03-23
 
 - Install updates automatically when Outlook closes without force-quitting